added further user management system tests
